requirements_client.txt
argcomplete==1.0.0
beautifulsoup4==4.4.1
cffi==1.4.2
chardet==2.3.0
cryptography==1.1.2
EbookLib==0.15
enum34==1.1.2
idna==2.0
ipaddress==1.0.16
lxml==3.5.0
pdfminer==20140328
Pillow==3.0.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
PyKCS11==1.3.0
python-docx==0.8.5
python-pptx==0.5.8
requests==2.9.1
six==1.10.0
SpeechRecognition==3.1.3
SQLAlchemy==1.0.11
textract==1.4.0
wheel==0.24.0
xlrd==0.9.4
XlsxWriter==0.7.7
